run OsmParser on a separate thread, docs updated
