+++ src/Viewer3D/CityMapGeometry.cc
@@ -26,10 +26,6 @@ void CityMapGeometry::setModelName(QString modelName)
 
 void CityMapGeometry::setOsmFilePath(QVariant value)
 {
-    if(_osmFilePath.compare(value.toString()) == 0){
-        return;
-    }
-
     clearViewer();
     _mapLoadedFlag = 0;
     _osmFilePath = value.toString();
@@ -51,16 +47,12 @@ void CityMapGeometry::setOsmParser(OsmParser *newOsmParser)
 
 bool CityMapGeometry::loadOsmMap()
 {
-    if(_mapLoadedFlag){
-        return true;
-    }
-
     if(!_osmParser){
         return false;
     }
-    _mapLoadedFlag = 1;
+
     _osmParser->parseOsmFile(_osmFilePath);
-    return true;
+    return false;
 }
 
 void CityMapGeometry::updateViewer()
@@ -71,7 +63,7 @@ void CityMapGeometry::updateViewer()
         return;
     }
 
-    if(loadOsmMap()){
+    if(_osmParser->mapLoaded()){
         _vertexData = _osmParser->buildingToMesh();
 
         int stride = 3 * sizeof(float);
