+++ src/Viewer3D/OsmParser.h
@@ -9,45 +9,13 @@
 #include <QVariant>
 #include <QDomElement>
 
+#include "OsmParserThread.h"
 ///     @author Omid Esrafilian <esrafilian.omid@gmail.com>
 
 class Viewer3DSettings;
 
 class OsmParser : public QObject
 {
-    struct BuildingType
-    {
-        std::vector<QGeoCoordinate> points_gps;
-        std::vector<QGeoCoordinate> points_gps_inner;
-        std::vector<QVector2D> points_local;
-        std::vector<QVector2D> points_local_inner;
-        std::vector<QVector3D> triangulated_mesh;
-        QVector2D bb_max = QVector2D(-1e6, -1e6); //bounding boxes
-        QVector2D bb_min = QVector2D(1e6, 1e6); //bounding boxes
-        float height;
-        float levels;
-
-        void append(std::vector<QGeoCoordinate> newPoints, bool isInner){
-            for(uint i=0; i<newPoints.size(); i++){
-                if(isInner){
-                points_gps_inner.push_back(newPoints[i]);
-                }else{
-                    points_gps.push_back(newPoints[i]);
-                }
-            }
-        }
-
-        void append(std::vector<QVector2D> newPoints, bool isInner){
-            for(uint i=0; i<newPoints.size(); i++){
-                if(isInner){
-                    points_local_inner.push_back(newPoints[i]);
-                }else{
-                    points_local.push_back(newPoints[i]);
-                }
-            }
-        }
-    };
-
     Q_OBJECT
 
     // Q_PROPERTY(float buildingLevelHeight READ buildingLevelHeight WRITE setBuildingLevelHeight NOTIFY buildingLevelHeightChanged)
@@ -62,21 +30,17 @@ public:
 
     float buildingLevelHeight(void){return _buildingLevelHeight;}
     void parseOsmFile(QString filePath);
-    void decodeNodeTags(QDomElement& xmlComponent, QMap<uint64_t, QGeoCoordinate> &nodeMap);
-    void decodeBuildings(QDomElement& xmlComponent, QMap<uint64_t, BuildingType > &buildingMap, QMap<uint64_t, QGeoCoordinate> &nodeMap, QGeoCoordinate gpsRef);
-    void decodeRelations(QDomElement& xmlComponent, QMap<uint64_t, BuildingType > &buildingMap, QMap<uint64_t, QGeoCoordinate> &nodeMap, QGeoCoordinate gpsRef);
 
     QByteArray buildingToMesh();
 
     void trianglateWallsExtrudedPolygon(std::vector<QVector3D>& triangulatedMesh, std::vector<QVector2D> verticesCcw, float h, bool inverseOrder=0, bool duplicateStartEndPoint=0);
     void trianglateRectangle(std::vector<QVector3D>& triangulatedMesh, std::vector<QVector3D> verticesCcw, bool invertNormal);
-    std::pair<QGeoCoordinate, QGeoCoordinate> getMapBoundingBoxCoordinate(){ return std::pair(_coordinate_min, _coordinate_max);}
+    std::pair<QGeoCoordinate, QGeoCoordinate> getMapBoundingBoxCoordinate(){ return std::pair(_coordinateMin, _coordinateMax);}
 
 private:
+    OsmParserThread* _osmParserWorker;
     QGeoCoordinate _gpsRefPoint;
-    QMap<uint64_t, QGeoCoordinate> _mapNodes;
-    QMap<uint64_t, BuildingType> _mapBuildings;
-    QGeoCoordinate _coordinate_min, _coordinate_max; //Osm map bounding boxes in global coordinate
+    QGeoCoordinate _coordinateMin, _coordinateMax; //Osm map bounding boxes in global coordinate
 
 
     bool _gpsRefSet;
@@ -94,6 +58,8 @@ signals:
 
 private slots:
     void setBuildingLevelHeight(QVariant value);
+    void osmParserFinished(bool isValid);
+
 
 };
